hide ctor of phone for safety
